Fix migrating False boolean values, handling of "unset" values:
If an AT schema had a `BooleanField` that defaults to `True` and an instance
had a value set to `False`, the migration would take that `False` to mean to
leave the value as unset on the new DX instance which would then get the
default `True` value.  Fix this bug by properly handling unset values by
checking for `AttributeError`.